feat: dynamic sdk config
