+++ deployments/dapps/create-react-app/src/components/ItemView.tsx
