+++ deployments/dapps/create-react-app/src/components/layout.tsx
@@ -1,0 +1,19 @@
+import React, { ReactNode } from 'react';
+import SDKConfig from './sdk-config';
+
+export interface LayoutProps {
+  children: ReactNode;
+}
+
+export const Layout = ({ children }: LayoutProps) => {
+  return (
+    <div>
+      <SDKConfig
+        onHomePress={() => {
+          window.location.href = '/';
+        }}
+      />
+      {children}
+    </div>
+  );
+};
