+++ packages/devnext/package.json
@@ -32,6 +32,7 @@
     "bufferutil": "^4.0.7",
     "encoding": "^0.1.13",
     "ethers": "5.7.2",
+    "next-qrcode": "^2.5.1",
     "preact": "^10.13.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
