+++ packages/devnext/src/components/package-version.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import packageJson from '../../package.json';
+
+export default function PackageVersion() {
+  return <div>PackageVersion: {packageJson.version}</div>;
+}
