+++ packages/devnext/src/pages/_app.tsx
@@ -1,19 +1,31 @@
-import { MetaMaskUIProvider } from '@metamask/sdk-react-ui';
-import React from 'react';
-import type { AppProps } from 'next/app';
-import '../styles/globals.css';
 import { config } from '@fortawesome/fontawesome-svg-core';
 import '@fortawesome/fontawesome-svg-core/styles.css';
+import { MetaMaskUIProvider } from '@metamask/sdk-react-ui';
+import type { AppProps } from 'next/app';
+import React from 'react';
+import {
+  SDKConfigProvider,
+  useSDKConfig,
+} from '../providers/sdkconfig-context';
+import '../styles/globals.css';
+import { Layout } from '../components/layout';
 config.autoAddCss = false;
 
-export default function App({ Component, pageProps }: AppProps) {
+const WithSDKConfig = ({ children }: { children: React.ReactNode }) => {
+  const {
+    socketServer,
+    infuraAPIKey,
+    useDeeplink,
+    checkInstallationImmediately,
+  } = useSDKConfig();
+
   return (
     <MetaMaskUIProvider
       debug={true}
       sdkOptions={{
-        communicationServerUrl: process.env.NEXT_PUBLIC_COMM_SERVER_URL,
+        communicationServerUrl: socketServer,
         enableDebug: true,
-        infuraAPIKey: process.env.NEXT_PUBLIC_INFURA_API_KEY,
+        infuraAPIKey,
         readonlyRPCMap: {
           '0x539': process.env.NEXT_PUBLIC_PROVIDER_RPCURL ?? '',
         },
@@ -24,8 +36,8 @@ export default function App({ Component, pageProps }: AppProps) {
           serviceLayer: false,
           plaintext: true,
         },
-        useDeeplink: false,
-        checkInstallationImmediately: false,
+        useDeeplink,
+        checkInstallationImmediately,
         storage: {
           enabled: true,
         },
@@ -38,7 +50,19 @@ export default function App({ Component, pageProps }: AppProps) {
         },
       }}
     >
-      <Component {...pageProps} />
+      {children}
     </MetaMaskUIProvider>
+  );
+};
+
+export default function App({ Component, pageProps }: AppProps) {
+  return (
+    <SDKConfigProvider>
+      <WithSDKConfig>
+        <Layout>
+          <Component {...pageProps} />
+        </Layout>
+      </WithSDKConfig>
+    </SDKConfigProvider>
   );
 }
